[Behat] Extract product attribute form element from create and update pages
