+++ src/Sylius/Behat/Page/Admin/ProductAttribute/CreatePageInterface.php
@@ -13,29 +13,8 @@ declare(strict_types=1);
 
 namespace Sylius\Behat\Page\Admin\ProductAttribute;
 
-use Behat\Mink\Exception\ElementNotFoundException;
 use Sylius\Behat\Page\Admin\Crud\CreatePageInterface as BaseCreatePageInterface;
 
 interface CreatePageInterface extends BaseCreatePageInterface
 {
-    public function specifyCode(string $code): void;
-
-    public function nameIt(string $name, string $language): void;
-
-    public function disableTranslatability(): void;
-
-    public function isTypeDisabled(): bool;
-
-    public function addAttributeValue(string $value, string $localeCode): void;
-
-    public function specifyMinValue(int $min): void;
-
-    public function specifyMaxValue(int $max): void;
-
-    public function checkMultiple(): void;
-
-    /**
-     * @throws ElementNotFoundException
-     */
-    public function getValidationErrors(): string;
 }
