+++ src/Sylius/Behat/Resources/config/services/elements/admin/product.xml
@@ -25,6 +25,14 @@
         </service>
 
         <service
+            id="sylius.behat.element.admin.product_attribute.form"
+            class="Sylius\Behat\Element\Admin\ProductAttribute\FormElement"
+            parent="sylius.behat.element.admin.crud.form"
+        >
+            <argument type="service" id="Sylius\Behat\Service\Helper\AutocompleteHelperInterface" />
+        </service>
+
+        <service
             id="sylius.behat.element.admin.product.association_form"
             class="Sylius\Behat\Element\Admin\Product\AssociationsFormElement"
             parent="sylius.behat.element.admin.crud.form"
